requirements.txt
python-telegram-bot
pymongo
python-dotenv
os
gunicorn